minor updates
Updated packages install list for rhel to include sssd and samba-tools-common as those seem to be needed for the realm command to work. Removed duplicate settings for dns resolver for Rhel and Ubuntu and limited the setting for this so that it simply and correctly is set in /etc/resolv.conf. Added a note that the domain controller will need to be able to resolve the Linux server for it's SPN's to function properly.
Put a note at the top pre-requisites so that the hostname of the linux server is resolvable on the domain controllers, as some customers might not have that.
Added an update to the ktutil command so that it'll be more clear what records users should prune from /etc/krb5.keytab

Hold off on posting this until the keytab configuration is confirmed.